Move JsonConverter, Validator into utils/
